requirements.txt
customtkinter==5.2.2
Pillow==10.4.0
customtkinter>=5.2.2
pillow>=10.3.0


